requirements.txt
django
django-crispy-forms
django-mathfilters
django-credit-cards
